update views for backend
