+++ backend/views/file.py
@@ -1,0 +1,38 @@
+from database import db_session, engine
+from flask import jsonify, request
+from models.file import File
+from utils import app, validate_jwt
+from worker import scrape_data_from_google
+
+
+@app.route('/csv', methods=['GET', 'POST'])
+def process_csv():
+    token_result = validate_jwt(request.headers.get('Authorization'))
+    if token_result == 401:
+        return "Unauthorized", 401
+
+    if request.method == 'GET':
+        result = engine.execute("""
+                SELECT 
+                    id, 
+                    filename, 
+                    keywords,
+                    created,
+                    (SELECT COUNT(*) >= f.keywords FROM data WHERE file_id=f.id) AS status
+                FROM file f 
+                WHERE user_id=%s
+                ORDER BY created DESC;
+            """, [token_result["sub"]])
+        return jsonify([list(row) for row in result]), 200
+    elif request.method == 'POST':
+        request_body = request.json
+        new_file = File(
+            user_id = token_result["sub"],
+            filename = request_body["filename"],
+            keywords = len(request_body["keywords"])
+        )
+        db_session.add(new_file)
+        db_session.commit()
+        for keyword in request_body["keywords"]:
+            scrape_data_from_google.apply_async(args=[new_file.id, keyword])
+        return "Upload Completed", 200
